Constrain :runtime_toolchain_type toolchains by target platform
The toolchain type @bazel_tools//tools/jdk:runtime_toolchain_type is for
a JDK to be executed on the target platform e.g. to run Java bytecode.
The toolchain type @bazel_tools//tools/jdk:toolchain_type is for a JDK
to be executed on the build platform e.g. to compiled Java source
code. Therefore, a toolchain of type
@bazel_tools//tools/jdk:runtime_toolchain_type should be constrained by
target_compatible_with, not exec_compatible_with.